requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.2
bootstrap4==0.1.0
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==2.2.8
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-registration==3.0.1
django-tinymce==2.8.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-decouple==3.3
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
typed-ast==1.4.0
whitenoise==5.0.1
wrapt==1.11.2
